backend/authentication/requirement_home.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
psycopg2-binary
Django
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyotp==2.9.0
qrcode==8.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.2.0
simplejwt==2.0.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.1